Move Response extension functions in MyanimelistApi class to a file
- This is to prepare for the next commit, adding a random manga feature to the app.
- consumeBody() will be called from another place which is why moving it to an accessible file is necessary.
- Placed the extension functions into the newly created OkHttpResponseExtensions file.
